Admin | user | non user for NavItem
I have two roles: "user" and "admin." If a user is logged in, they should be able to access navigation items with IDs 1, 2, 3, and 4, but not item 5 (which is reserved for admins). Similarly, even if a user is not logged in, they should still be able to see navigation items with IDs 1, 2, 3, and 4, while item 5 remains exclusive to admin access. When an admin is logged in, they have unrestricted access to all navigation items, including items with IDs 1, 2, 3, 4, and 5.