feat: Add Login Redirection Feature : Admin
Changes include:
- A new context state variable `prevPath` has been added to track the previous route.
- The `ContextProvider` has been updated to include `prevPath` and its setter function `setPrevPath`.
- A new `RouteTracker` component has been added to update `prevPath` whenever the route changes.
- The `IfLoginRedirect` component has been modified to use `prevPath` from the context to redirect the user after login.

With these changes, when a user logs in, they will be redirected to the previous page they were on, unless they directly accessed the login page, in which case they will be redirected to the home page. This provides a more intuitive and seamless navigation experience for the user.